fix up dequeue stage
