+++ projects/Plasma MVP/JavaScript/Dequeue/completionMessage.md
@@ -1,1 +1,5 @@
-completed
+## DQ Successful!
+
+![Dairy Queen](https://res.cloudinary.com/divzjiip8/image/upload/v1553131689/2000px-Dairy_Queen_logo_tczqyf.webp)
+
+Well done. Now we use our Exit Queue for great justice!
